+++ client/src/components/fontComponent/FontPreview.jsx
@@ -1,13 +1,19 @@
 import fontWeightConversion from "../../utils/fontWeightConversion";
+import fontWeightMap from "../../data/fontweight";
 
 function FontPreview({ fontFamily, fontWeight, className }) {
   return (
     <div
-      className={`h-auto min-h-[110px] flex justify-center  text-lightwhite overflow-hidden ${className}`}
+      className={`h-auto min-h-[110px] my-4 border-b border-lightwhite flex justify-center  text-lightwhite overflow-hidden ${className}`}
     >
       <div className="w-full">
+        <span className="text-xl text-lightwhite">
+          {fontWeightMap[fontWeightConversion(fontWeight)]} /{" "}
+          {fontWeightConversion(fontWeight)}
+        </span>
+
         <h1
-          className={`my-5 text-7xl outline-none`}
+          className={`my-5 text-9xl outline-none`}
           suppressContentEditableWarning={true}
           contentEditable={true}
           spellCheck={false}
