+++ client/src/components/index.jsx
@@ -2,6 +2,6 @@ import Header from "./Header";
 import FontAbout from "./FontAbout";
 import FontLayout from "./FontLayout";
 import Footer from "./Footer";
-import Font from "./fontComponent";
+import Font from "./Font";
 
 export { Header, Font, FontAbout, FontLayout, Footer };
