+++ client/src/pages/Home.jsx
@@ -1,11 +1,29 @@
+import { useEffect } from "react";
+
 import { Font, FontAbout, FontLayout } from "../components";
+import { generateRandomNum } from "../utils";
+
+import useFont from "../hooks/useFont";
 
 function Home() {
+  const { fetchAllFonts, dispatch } = useFont();
+
+  useEffect(() => {
+    (async () => {
+      const data = await fetchAllFonts();
+      dispatch({ type: "SET_FONTS_LIST", payload: data });
+
+      dispatch({
+        type: "SET_CURRENT_FONT",
+        payload: data[generateRandomNum(data.length)],
+      });
+    })();
+  }, []);
+
   return (
     <div className="h-auto w-full bg-black">
-      {/* hello */}
       <Font />
-      {/* About the font section */}
+      <FontLayout />
       <FontAbout />
     </div>
   );
